refactor(agent): honor shared server boundaries
