+++ apps/server/api/src/collections/agent-threads/agent-threads.module.ts
@@ -6,9 +6,10 @@
 
 import { AgentMessagesModule } from '@api/collections/agent-messages/agent-messages.module';
 import { AgentThreadsController } from '@api/collections/agent-threads/controllers/agent-threads.controller';
-import { AgentScopeContextService } from '@api/collections/agent-threads/services/agent-scope-context.service';
 import { AgentThreadsService } from '@api/collections/agent-threads/services/agent-threads.service';
 import { UsersModule } from '@api/collections/users/users.module';
+import { PrismaService } from '@api/shared/modules/prisma/prisma.service';
+import { AgentScopeContextService, SERVER_TOKENS } from '@genfeedai/server';
 import { forwardRef, Module } from '@nestjs/common';
 
 @Module({
@@ -18,6 +19,10 @@ import { forwardRef, Module } from '@nestjs/common';
     forwardRef(() => AgentMessagesModule),
     forwardRef(() => UsersModule),
   ],
-  providers: [AgentScopeContextService, AgentThreadsService],
+  providers: [
+    AgentScopeContextService,
+    AgentThreadsService,
+    { provide: SERVER_TOKENS.prisma, useExisting: PrismaService },
+  ],
 })
 export class AgentThreadsModule {}
