+++ apps/server/workers/src/crons/posts/cron.posts.module.ts
@@ -1,5 +1,4 @@
 import { ActivitiesModule } from '@api/collections/activities/activities.module';
-import { AgentThreadsModule } from '@api/collections/agent-threads/agent-threads.module';
 import { CredentialsModule } from '@api/collections/credentials/credentials.module';
 import { OrganizationsModule } from '@api/collections/organizations/organizations.module';
 import { PostsModule } from '@api/collections/posts/posts.module';
@@ -7,23 +6,31 @@ import { SystemWorkflowProvenanceService } from '@api/collections/workflows/serv
 import { PublishersModule } from '@api/services/integrations/publishers/publishers.module';
 import { QuotaModule } from '@api/services/quota/quota.module';
 import { WebhookClientModule } from '@api/services/webhook-client/webhook-client.module';
+import { AgentScopeContextService, SERVER_TOKENS } from '@genfeedai/server';
+import { PrismaModule } from '@libs/prisma/prisma.module';
+import { PrismaService } from '@libs/prisma/prisma.service';
 import { forwardRef, Module } from '@nestjs/common';
 import { CronPostsService } from '@workers/crons/posts/cron.posts.service';
 import { WorkersQueuesModule } from '@workers/queues/queues.module';
 
 @Module({
   imports: [
-    forwardRef(() => AgentThreadsModule),
     forwardRef(() => ActivitiesModule),
     forwardRef(() => CredentialsModule),
     forwardRef(() => OrganizationsModule),
     forwardRef(() => PostsModule),
     forwardRef(() => WebhookClientModule),
     PublishersModule,
+    PrismaModule,
     QuotaModule,
     forwardRef(() => WorkersQueuesModule),
   ],
   exports: [CronPostsService],
-  providers: [CronPostsService, SystemWorkflowProvenanceService],
+  providers: [
+    AgentScopeContextService,
+    CronPostsService,
+    SystemWorkflowProvenanceService,
+    { provide: SERVER_TOKENS.prisma, useExisting: PrismaService },
+  ],
 })
 export class CronPostsModule {}
